Move files into different packages
