+++ src/main/java/com/Duke/Commands/BlahCommand.java
@@ -1,7 +1,7 @@
 package com.Duke.Commands;
 
-import com.Duke.ProcessManager.DukeException;
-import com.Duke.ProcessManager.UI;
+import com.Duke.TaskManager.DukeException;
+import com.Duke.TaskManager.UI;
 
 public class BlahCommand extends Command{
