+++ src/main/java/com/Duke/Commands/DeleteCommand.java
@@ -1,8 +1,8 @@
 package com.Duke.Commands;
 
-import com.Duke.ProcessManager.DukeException;
-import com.Duke.ProcessManager.TaskList;
-import com.Duke.ProcessManager.UI;
+import com.Duke.TaskManager.DukeException;
+import com.Duke.TaskManager.TaskList;
+import com.Duke.TaskManager.UI;
 
 public class DeleteCommand extends Command{
     private final TaskList ls;
