+++ src/main/java/com/Duke/Commands/EventCommand.java
@@ -1,7 +1,7 @@
 package com.Duke.Commands;
 
-import com.Duke.ProcessManager.TaskList;
-import com.Duke.ProcessManager.UI;
+import com.Duke.TaskManager.TaskList;
+import com.Duke.TaskManager.UI;
 import com.Duke.Tasks.Event;
 
 import java.time.LocalTime;
