+++ src/main/java/com/Duke/TaskManager/DukeParser.java
@@ -1,4 +1,4 @@
-package com.Duke.ProcessManager;
+package com.Duke.TaskManager;
 
 import com.Duke.Commands.*;
