+++ src/main/java/com/Duke/TaskManager/TaskList.java
@@ -1,4 +1,4 @@
-package com.Duke.ProcessManager;
+package com.Duke.TaskManager;
 
 import com.Duke.Tasks.Task;
