+++ src/main/java/com/Duke/Tasks/Deadline.java
@@ -1,6 +1,6 @@
 package com.Duke.Tasks;
 
-import com.Duke.ProcessManager.DukeException;
+import com.Duke.TaskManager.DukeException;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
