+++ src/main/java/com/Duke/Tasks/Event.java
@@ -1,6 +1,6 @@
 package com.Duke.Tasks;
 
-import com.Duke.ProcessManager.DukeException;
+import com.Duke.TaskManager.DukeException;
 
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
